Remove unsubscribe url and cast suppression group id to an integer

+++ src/email-providers/class-sendgrid.php
@@ -169,14 +169,13 @@ class Sendgrid implements Email_Provider {
 		$request_body->categories   = [];
 		$request_body->email_config = (object) [];
 		$request_body->send_to      = (object) [];
-		// $request_body->email_config->custom_unsubscribe_url = '';
 		$request_body->email_config->html_content           = $html_content;
 		$request_body->email_config->ip_pool                = null;
 		$request_body->email_config->plain_content          = $text_content;
 		$request_body->email_config->generate_plain_content = true;
 		$request_body->email_config->sender_id              = $sender_id ?? 0;
 		$request_body->email_config->subject                = $subject;
-		$request_body->email_config->suppression_group_id   = get_post_meta( $newsletter_id, 'nb_newsletter_suppression_group', true );
+		$request_body->email_config->suppression_group_id   = (int) get_post_meta( $newsletter_id, 'nb_newsletter_suppression_group', true );
 
 		$request_body->send_to->list_ids    = $list_ids;
 		$request_body->send_to->segment_ids = [];
